We now directly create a BigDecimal from a given string instead of turning it into a double.
Otherwise we loose some accuracy (e.g. 0.1 becomes a very strange value to due numeric reasons).

+++ src/main/java/sirius/kernel/commons/Amount.java
@@ -94,7 +94,7 @@ public class Amount implements Comparable<Amount> {
         if (Strings.isEmpty(value)) {
             return NOTHING;
         }
-        return of(NLS.parseMachineString(Double.class, value));
+        return of(NLS.parseMachineString(BigDecimal.class, value));
     }
 
     /**
@@ -109,7 +109,7 @@ public class Amount implements Comparable<Amount> {
         if (Strings.isEmpty(value)) {
             return NOTHING;
         }
-        return of(NLS.parseUserString(Double.class, value));
+        return of(NLS.parseUserString(BigDecimal.class, value));
     }
 
     /**
